+++ src/api/build.ts
@@ -42,7 +42,6 @@ export async function build({
 	routes = path.resolve(cwd, routes);
 	output = path.resolve(cwd, output);
 	static_files = path.resolve(cwd, static_files);
-	dest = path.resolve(cwd, dest);
 
 	if (legacy && bundler === 'webpack') {
 		throw new Error(`Legacy builds are not supported for projects using webpack`);
@@ -119,10 +118,15 @@ export async function build({
 	let serviceworker_stats;
 
 	if (serviceworker) {
+
+		const client_files = client_result.chunks
+			.filter(chunk => !chunk.file.endsWith('.map')) // SW does not need to cache sourcemap files
+			.map(chunk => `client/${chunk.file}`);
+
 		create_serviceworker_manifest({
 			manifest_data,
 			output,
-			client_files: client_result.chunks.map(chunk => `client/${chunk.file}`),
+			client_files,
 			static_files
 		});