+++ test/apps/scroll/test.ts
@@ -31,7 +31,7 @@ describe('scroll', function() {
 		await start();
 
 		const scrollY = await page.evaluate(() => window.scrollY);
-		assert.ok(scrollY > 0, scrollY);
+		assert.ok(scrollY > 0, String(scrollY));
 	});
 
 	it('scrolls to any deeplink if it was already active', async () => {
@@ -39,17 +39,17 @@ describe('scroll', function() {
 		await start();
 
 		let scrollY = await page.evaluate(() => window.scrollY);
-		assert.ok(scrollY > 0, scrollY);
+		assert.ok(scrollY > 0, String(scrollY));
 
 		scrollY = await page.evaluate(() => {
 			window.scrollTo(0, 0)
 			return window.scrollY
 		});
-		assert.ok(scrollY === 0, scrollY);
+		assert.ok(scrollY === 0, String(scrollY));
 
 		await page.click('[href="tall-page#foo"]');
 		scrollY = await page.evaluate(() => window.scrollY);
-		assert.ok(scrollY > 0, scrollY);
+		assert.ok(scrollY > 0, String(scrollY));
 	});
 
 	it('resets scroll when a link is clicked', async () => {