+++ test/apps/with-sourcemaps-webpack/src/server.js
@@ -1,0 +1,8 @@
+import polka from 'polka';
+import * as sapper from '@sapper/server';
+
+const { PORT } = process.env;
+
+polka()
+	.use(sapper.middleware())
+	.listen(PORT);